ai_service/requirements.txt
fastapi
uvicorn[standard]
ultralytics
Pillow
python-multipart
numpy<2.0
